Move interpolation from .io to .monitoring
