+++ src/ctapipe/io/tableloader.py
@@ -13,8 +13,8 @@ from astropy.utils.decorators import lazyproperty
 
 from ..core import Component, Provenance, traits
 from ..instrument import FocalLengthKind, SubarrayDescription
+from ..monitoring.interpolation import PointingInterpolator
 from .astropy_helpers import join_allow_empty, read_table
-from .interpolation import PointingInterpolator
 
 __all__ = ["TableLoader"]
