+++ src/ctapipe/monitoring/__init__.py
@@ -1,0 +1,23 @@
+"""
+Module for handling monitoring data.
+"""
+from .aggregator import PlainAggregator, SigmaClippingAggregator, StatisticsAggregator
+from .interpolation import Interpolator, PointingInterpolator
+from .outlier import (
+    MedianOutlierDetector,
+    OutlierDetector,
+    RangeOutlierDetector,
+    StdOutlierDetector,
+)
+
+__all__ = [
+    "PlainAggregator",
+    "SigmaClippingAggregator",
+    "StatisticsAggregator",
+    "OutlierDetector",
+    "RangeOutlierDetector",
+    "MedianOutlierDetector",
+    "StdOutlierDetector",
+    "Interpolator",
+    "PointingInterpolator",
+]
