+++ src/ctapipe/monitoring/interpolation.py
